+++ src/conversions/osstr.rs
@@ -145,6 +145,12 @@ impl IntoPy<PyObject> for OsString {
     }
 }
 
+impl<'a> IntoPy<PyObject> for &'a OsString {
+    fn into_py(self, py: Python) -> PyObject {
+        self.to_object(py)
+    }
+}
+
 #[cfg(test)]
 mod test {
     use crate::{types::PyString, IntoPy, PyObject, Python, ToPyObject};
@@ -205,6 +211,7 @@ mod test {
             let os_str = OsStr::new("Hello\0\n🐍");
             test_roundtrip::<&OsStr>(py, os_str);
             test_roundtrip::<OsString>(py, os_str.to_os_string());
+            test_roundtrip::<&OsString>(py, &os_str.to_os_string());
         })
     }
 }